% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm2.R
\name{gmm2}
\alias{gmm2}
\title{Parameters Estimation of Mixture of Two Normal Distributions by EM Algorithms}
\usage{
gmm2(p, mu, sig, data)
}
\arguments{
\item{p}{the initial of the weights of two normal distributions}

\item{mu}{the initial of the population means of two normal distributions}

\item{sig}{the initial of the population deviations of two normal distributions}

\item{data}{the data which generated from the mixture of two normal distributions}
}
\value{
a list of estimated values of parameters
}
\description{
Estimate the parameters of mixture of two normal distribution by EM algorithms
}
\examples{
\dontrun{
n <- 200
u <- runif(n)
x <- ifelse(u<.7,rnorm(n),rnorm(3,9))
gmm2(c(0.65,0.35),c(-0.1,3.21),c(1.59,7.99),x)
}
}
