% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixnormal2.R
\name{mixnormal2}
\alias{mixnormal2}
\title{Predict the Density Value at certain pointsPoint From the Mixture of Two Normal Distributions by MLE}
\usage{
mixnormal2(x, initial, data)
}
\arguments{
\item{x}{the specific point or points from the mixture of two normal distributions}

\item{initial}{a vector consisting the initial values of weight, means and deviations of two normal distributions}

\item{data}{the data which generated from the mixture of two normal distributions}
}
\value{
an estimated value
}
\description{
Estimate the parameters of mixture of two normal distribution by MLE and then predict the density value at certain points
}
\examples{
\dontrun{
n <- 200
u <- runif(n)
x <- ifelse(u<.7,rnorm(n),rnorm(n,3,9))
mixnormal2(0.8,c(.65,-0.1,1.59,3.21,7.99),x)
}
}
