% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolisC}
\alias{rwMetropolisC}
\title{A Gibbs sampler using Rcpp}
\usage{
rwMetropolisC(x0, N, sigma)
}
\arguments{
\item{N}{the number of samples}

\item{thin}{the number of between-sample random numbers}
}
\value{
a random sample of size
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
rnC <- gibbsC(100,10)
par(mfrow=c(2,1));
plot(rnC[,1],type='l')
plot(rnC[,2],type='l')
}
}
